Update to address problems mentioned in review...
1. Removed details directory
2. Use 2 space indention
3. `.ixx` suffix for module
4. Merged `proxy_interface` into `proxy`
