--- include/proxy/proxy_interface.cppm
@@ -1,10 +1,0 @@
-module;
-
-#include <proxy/proxy.h>
-
-export module proxy_interface;
-
-export namespace pro::details {
-using details::adl_accessor_arg_t;
-using details::non_proxy_arg;
-} // namespace pro::details
